Move OnAnnotationPressListener to the other listeners
