+++ android-pdf-viewer/src/main/java/com/github/barteksc/pdfviewer/listener/OnAnnotationPressListener.kt
@@ -1,4 +1,4 @@
-package com.github.barteksc.pdfviewer.annotation.core.pdf
+package com.github.barteksc.pdfviewer.listener
 
 fun interface OnAnnotationPressListener {
     fun openDocumentation(schemaId: Long, documentId: String?)
