+++ sample/src/main/java/com/github/barteksc/sample/PDFViewActivity.java
@@ -36,7 +36,7 @@ import androidx.core.app.ActivityCompat;
 import androidx.core.content.ContextCompat;
 
 import com.github.barteksc.pdfviewer.PDFView;
-import com.github.barteksc.pdfviewer.annotation.core.pdf.OnAnnotationPressListener;
+import com.github.barteksc.pdfviewer.listener.OnAnnotationPressListener;
 import com.github.barteksc.pdfviewer.listener.OnErrorListener;
 import com.github.barteksc.pdfviewer.listener.OnLoadCompleteListener;
 import com.github.barteksc.pdfviewer.listener.OnLongPressListener;
